roll commit source on epoch change
